Packaged code
- created importable mom6_parameter_scanner module
- broke out CLI script as a stand-alone
- Python packaging/install scripts
