+++ mom6_parameter_scanner/MOM6param_scan.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 import argparse
 import collections
 import fnmatch
@@ -11,61 +9,9 @@ import shlex
 import tarfile
 import warnings
 
-# Globals
-debug = False
+__all__ = ["main","openParameterFile","Parameters","Namelists","Log","splitPath","uniq"]
 
-# For HTML
-html_header = '''<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-<html>
-<head>
-  <meta content="text/html; charset=ISO-8859-1"
- http-equiv="content-type">
-  <title>MOM6 parameters</title>
-</head>
-'''
-
-
-def parseCommandLine():
-  """
-  Parse the command line positional and optional arguments.
-  This is the highest level procedure invoked from the very end of the script.
-  """
-  global debug # Declared global in order to set it
-
-  # Arguments
-  parser = argparse.ArgumentParser(description=
-      '''
-      MOM6param_scan.py scans MOM_parameter_doc files and summarizes configurations.
-      It can also parse input.nml and logfiles for Fortran namelists.
-      ''',
-      epilog='Written by A.Adcroft, 2015.')
-  parser.add_argument('files', type=str, nargs='+',
-      help='''parameter-files/tar-files to scan.''')
-  parser.add_argument('-nml','--namelist', action='store_true',
-      help='scan for Fortran namelists.')
-  parser.add_argument('-mnml','--mom6namelist', action='store_true',
-      help='scan for Fortran namelists ignoring logfile.000000.out.')
-  parser.add_argument('-log','--log', action='store_true',
-      help='scan for model output in FMS log file.')
-  parser.add_argument('-if','--ignore_files', action='append', default=[],
-      help='file patterns to ignore when searching.')
-  parser.add_argument('-m','--assume_mom6', action='store_true',
-      help='assume FILE is a MOM6 parameter file. By default only MOM_parameter_doc.{all,short} are scanned.')
-  parser.add_argument('-x','--exclude', action='append', default=[],
-      help='parameters to exclude.')
-  parser.add_argument('-fmt','--format', choices=['json', 'html', 'text'], default='json',
-      help='output format.')
-  parser.add_argument('-t','--transpose', action='store_true',
-      help='transpose html table.')
-  parser.add_argument('-d','--debug', action='store_true',
-      help='turn on debugging information.')
-  args = parser.parse_args()
-
-  debug = args.debug
-
-  main(args)
-
-def main(args):
+def main(args,debug=False):
   """
   Does the actual work
   """
@@ -146,7 +92,7 @@ def main(args):
           for e in table.keys(): print(table[e][p],end=" ")
   return P
 
-def openParameterFile(file,parameter_files=['*MOM_parameter_doc.all', '*MOM_parameter_doc.short', 'MOM_input'],ignore_files=[]):
+def openParameterFile(file,parameter_files=['*MOM_parameter_doc.all', '*MOM_parameter_doc.short', 'MOM_input'],ignore_files=[],debug=False):
   """Returns python file object for a MOM_parameter_doc file, that might be inside a tar-file."""
 
   # Check if file is a directory to search
@@ -359,6 +305,3 @@ def uniq(list):
   for e in list:
     if not e in newlist: newlist.append(e)
   return newlist
-
-# Invoke parseCommandLine(), the top-level prodedure
-if __name__ == '__main__': parseCommandLine()
