+++ mom6_parameter_scanner/__init__.py
@@ -1,0 +1,4 @@
+""" mom6_parameter_scanner """
+
+from .MOM6param_scan import main, openParameterFile, Parameters, Namelists, Log, splitPath, uniq
+
