+++ setup.cfg
@@ -1,0 +1,6 @@
+[options]
+packages = mom6_parameter_scanner
+zip_safe = False  # https://mypy.readthedocs.io/en/latest/installed_packages.html
+include_package_data = True
+python_requires = >=3.8
+
